backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
spacy==3.7.6
PyMuPDF==1.24.10
python-docx==1.1.2
openai==1.51.0
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
reportlab==4.2.4
groq==0.11.0
httpx==0.27.2
google-auth==2.35.0
